refactor(Participant) : 참가자는 State 만 가지고 있게 변경
- GamePlayer : PlayerInfo 를 가짐
- Dealer : Name 만 가짐

- 이에 따른 테스트 코드 변경
