+++ src/main/java/blackjack/domain/player/Dealer.java
@@ -4,19 +4,22 @@ import blackjack.domain.card.Card;
 import blackjack.domain.card.Cards;
 
 public class Dealer extends Participant {
+    private final Name name;
     public static final Integer RECEIVE_SIZE = 16;
     private static final String DEFAULT_DEALER_NAME = "딜러";
 
     public Dealer(final Name name, final Cards cards) {
-        super(name, cards);
-    }
-
-    public Dealer(final PlayerInfo playerInfo) {
-        super(playerInfo);
+        super(cards);
+        this.name = name;
     }
 
     public static Dealer createDefaultDealer(final Cards cards) {
         return new Dealer(new Name(DEFAULT_DEALER_NAME), cards);
+    }
+
+    @Override
+    public boolean isReceivable() {
+        return this.state.calculate() <= RECEIVE_SIZE;
     }
 
     public Card getFirstCard() {
@@ -25,7 +28,11 @@ public class Dealer extends Participant {
     }
 
     @Override
-    public boolean isReceivable() {
-        return this.state.calculate() <= RECEIVE_SIZE;
+    public String getNameAsString() {
+        return name.asString();
+    }
+
+    public Name getName() {
+        return this.name;
     }
 }
