+++ src/main/java/blackjack/domain/player/GamePlayer.java
@@ -3,17 +3,51 @@ package blackjack.domain.player;
 
 import blackjack.domain.card.Cards;
 
+import java.util.Objects;
+
 public class GamePlayer extends Participant {
+    private final PlayerInfo playerInfo;
+    private static final BettingMoney DEFAULT_BETTING_MONEY = new BettingMoney(1000);
+
     public GamePlayer(final Name name, final Cards cards) {
-        super(name, cards);
+        super(cards);
+        this.playerInfo = new PlayerInfo(name, DEFAULT_BETTING_MONEY);
     }
 
     public GamePlayer(final PlayerInfo playerInfo) {
-        super(playerInfo);
+        super();
+        this.playerInfo = playerInfo;
+    }
+
+    @Override
+    public String getNameAsString() {
+        return this.playerInfo.name()
+                              .asString();
+    }
+
+    public int getBettingMoney() {
+        return this.playerInfo.bettingMoney()
+                              .value();
+    }
+
+    public Name getName() {
+        return this.playerInfo.name();
     }
 
     @Override
     public boolean isReceivable() {
         return this.state.isHit();
+    }
+
+    @Override
+    public boolean equals(final Object object) {
+        if (this == object) return true;
+        if (!(object instanceof final GamePlayer that)) return false;
+        return Objects.equals(this.playerInfo.name(), that.playerInfo.name());
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(this.playerInfo.name());
     }
 }
